Minor refactoring of InformationDisplay: - remove useless class data - do not recalculate layout on every frame draw; do it when size changed only - use unique_ptr instead of shared_ptr - remove useless includes

+++ gui/element.cpp
@@ -86,7 +86,7 @@ namespace gui
 
   void Element::draw(graphics::OverlayRenderer *r, math::Matrix<double, 3, 3> const & m) const
   {
-    for (unsigned i = 0; i < boundRects().size(); ++i)
+    for (size_t i = 0; i < boundRects().size(); ++i)
       r->drawRectangle(boundRects()[i], color(state()), depth());
   }
